refactor(client): move signer migration into module
